# openapi/tabs.yaml
openapi: 3.1.0
info:
  title: Tabs API
  version: 0.1.0
  description: >
    Record movements of money and read back the active financial cycle.
    All money amounts are integer cents. All dates are YYYY-MM-DD.

paths:
  /api/v1/movements:
    post:
      operationId: createMovement
      summary: Record a movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementInput'
      responses:
        '201':
          description: Movement recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        '400':
          description: Invalid amount, direction, or date; unknown category; or
            category of the other direction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: listMovements
      summary: List one page of the active cycle's movements
      parameters:
        - name: page
          in: query
          required: false
          description: >
            The page to return, from 1. A page beyond the last returns the
            last page.
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: One page of the active cycle's movements, newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementPage'
        '400':
          description: page is not a positive integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/movements/{id}:
    put:
      operationId: updateMovement
      summary: Replace a movement's amount, direction, category, date, and note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementInput'
      responses:
        '200':
          description: Movement updated; its id and created_at are unchanged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        '400':
          description: Invalid id, amount, direction, or date; unknown
            category; or category of the other direction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Movement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteMovement
      summary: Delete a movement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Movement deleted
        '404':
          description: Movement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/categories:
    get:
      operationId: listCategories
      summary: List all categories
      responses:
        '200':
          description: Seeded and user-created categories, in sort order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createCategory
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Blank name or invalid direction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: A category with that name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/cycles/current:
    get:
      operationId: getCurrentCycle
      summary: Active cycle bounds and balance
      responses:
        '200':
          description: >
            The active cycle. ends_on is exclusive: the cycle covers
            [starts_on, ends_on).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentCycle'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/dashboard/widgets:
    get:
      operationId: getWidgetSettings
      summary: Read the stored dashboard widget selection
      responses:
        '200':
          description: >
            The stored selection, or the all-enabled defaults when the
            user has never saved one. Unknown stored ids are dropped and
            missing ids fall back to their default.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetSettings'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: putWidgetSettings
      summary: Replace the whole dashboard widget selection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WidgetSettings'
      responses:
        '200':
          description: Selection stored; the response echoes it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetSettings'
        '400':
          description: >
            The mapping has an unknown id, a missing id, or a value that
            is not a boolean; the stored selection is unchanged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/habit:
    get:
      operationId: getHabit
      summary: Logging streak, XP and level for the active cycle
      description: >
        Derived on every request from the movements that exist now. A day
        counts as logged when a movement was created on it (in the configured
        time zone), whatever date the movement carries.
      responses:
        '200':
          description: The habit, or no habit at all on failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Habit'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Movement:
      type: object
      required: [id, amount_cents, direction, category_id, occurred_on, note, created_at]
      properties:
        id:
          type: integer
          format: int64
        amount_cents:
          type: integer
          format: int64
          description: Positive amount, in cents
        direction:
          type: string
          enum: [in, out]
        category_id:
          type: integer
          format: int64
        occurred_on:
          type: string
          format: date
        note:
          type: string
        created_at:
          type: string
          format: date-time

    MovementPage:
      type: object
      required: [items, page, total_pages, total]
      properties:
        items:
          type: array
          description: Up to 6 movements, by date then most recently recorded
          items:
            $ref: '#/components/schemas/Movement'
        page:
          type: integer
          description: The page returned, from 1
        total_pages:
          type: integer
          description: At least 1, also for an empty cycle
        total:
          type: integer
          description: Movements in the active cycle

    MovementInput:
      type: object
      required: [amount_cents, direction, category_id, occurred_on]
      properties:
        amount_cents:
          type: integer
          format: int64
          description: Positive amount, in cents
        direction:
          type: string
          enum: [in, out]
        category_id:
          type: integer
          format: int64
        occurred_on:
          type: string
          format: date
        note:
          type: string
          default: ''

    Category:
      type: object
      required: [id, name, direction, sort_order]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        direction:
          type: string
          enum: [in, out]
          description: Only movements with this direction may use the category
        sort_order:
          type: integer

    CategoryInput:
      type: object
      required: [name, direction]
      properties:
        name:
          type: string
        direction:
          type: string
          enum: [in, out]

    CurrentCycle:
      type: object
      required: [starts_on, ends_on, balance, category_totals]
      properties:
        starts_on:
          type: string
          format: date
          description: First day of the cycle, inclusive
        ends_on:
          type: string
          format: date
          description: First day of the next cycle, exclusive
        balance:
          $ref: '#/components/schemas/Balance'
        category_totals:
          type: array
          description: >
            One entry per category with a movement in the cycle: in before
            out, then by descending total, then by category id.
          items:
            $ref: '#/components/schemas/CategoryTotal'

    CategoryTotal:
      type: object
      required: [category_id, direction, total_cents, movement_count]
      properties:
        category_id:
          type: integer
          format: int64
        direction:
          type: string
          enum: [in, out]
        total_cents:
          type: integer
          format: int64
          description: Sum of the category's movements in the cycle, in cents
        movement_count:
          type: integer
          description: Movements of the category in the cycle

    Balance:
      type: object
      required: [total_in, total_out, net]
      properties:
        total_in:
          type: integer
          format: int64
          description: Sum of in movements, in cents
        total_out:
          type: integer
          format: int64
          description: Sum of out movements, in cents
        net:
          type: integer
          format: int64
          description: total_in minus total_out, in cents

    WidgetSettings:
      type: object
      required: [widgets]
      properties:
        widgets:
          type: object
          description: >
            Widget id to enabled. PUT must carry exactly these four ids;
            the server rejects any other mapping. The ids mirror the Go
            dashboard domain catalog — changing one means changing both.
          required:
            [net-balance, total-income, total-expenses, category-distribution]
          properties:
            net-balance:
              type: boolean
            total-income:
              type: boolean
            total-expenses:
              type: boolean
            category-distribution:
              type: boolean
          additionalProperties: false

    Habit:
      type: object
      required: [streak, days, xp, movement_xp, rules]
      properties:
        streak:
          type: object
          required: [current, today_logged]
          properties:
            current:
              type: integer
              description: >
                Consecutive logged days ending today, or yesterday when
                today is not yet logged. Counts across cycle boundaries.
            today_logged:
              type: boolean
        days:
          type: array
          description: Every day of the active cycle, first to last
          items:
            type: object
            required: [date, status]
            properties:
              date:
                type: string
                format: date
              status:
                type: string
                enum: [logged, missed, today_logged, today_pending, ahead]
        xp:
          type: object
          required: [total, level, level_starts_at, next_level_at]
          properties:
            total:
              type: integer
              description: XP over every existing movement, all cycles
            level:
              type: integer
            level_starts_at:
              type: integer
              description: Total XP at which the current level started
            next_level_at:
              type: integer
              description: Total XP at which the next level starts
        movement_xp:
          type: array
          description: XP earned by each movement dated in the active cycle
          items:
            type: object
            required: [movement_id, xp]
            properties:
              movement_id:
                type: integer
                format: int64
              xp:
                type: integer
        rules:
          type: array
          description: The XP rules the server applies
          items:
            type: object
            required: [id, label, xp]
            properties:
              id:
                type: string
              label:
                type: string
              xp:
                type: integer

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string